+++ src/procshared_malloc.cpp
@@ -21,9 +21,9 @@ procshared_malloc::~procshared_malloc()
 {
 }
 
-procshared_malloc::procshared_malloc( const char* p_shm_name, const char* p_id_dirname, off_t length, mode_t mode )
+procshared_malloc::procshared_malloc( const char* p_shm_name, const char* p_id_dirname, size_t length, mode_t mode )
 {
-	shm_obj_ = procshared_mem( p_shm_name, p_id_dirname, length, mode, [this]( void* p_mem, off_t len ) {
+	shm_obj_ = procshared_mem( p_shm_name, p_id_dirname, length, mode, [this]( void* p_mem, size_t len ) {
 		shm_heap_ = offset_malloc( p_mem, len );
 	} );
 }
